Refactor GameScreen into smaller widgets (#1148)
Split the in-game screen into focused components (map, overlays, side menu, dialogs) to reduce file size and improve maintainability without changing behavior.

Made-with: Cursor
