+++ app/lib/features/game/flame/game_screen_shared.dart
@@ -1,0 +1,11 @@
+import 'package:flutter/material.dart';
+
+/// Breakpoint for in-game narrow layout (side menu, reduced top bar). SPEC/ui/in-game-shell-narrow.md.
+const double kInGameNarrowBreakpoint = 600;
+
+/// Key for the base layer cycle button (for tests). SPEC/ui/empire-overview.md § Base layer display cycle.
+const Key kBaseLayerCycleButtonKey = Key('base_layer_cycle_button');
+
+/// Key for the home-to-capital button (for tests). SPEC/ui/empire-overview.md § Home-to-capital button.
+const Key kHomeToCapitalButtonKey = Key('home_to_capital_button');
+
