+++ app/lib/features/game/flame/victory_overlay.dart
@@ -1,0 +1,108 @@
+import 'package:flutter/material.dart';
+
+import 'package:colonizethis_models/colonizethis_models.dart' as ct_models;
+
+import '../../../widgets/ct_nine_patch_button.dart';
+import '../../../widgets/ct_panel.dart';
+
+/// Stateful overlay so "View final state" can hide the panel without a route (SPEC/game/victory.md).
+class VictoryOverlay extends StatefulWidget {
+  const VictoryOverlay({
+    required this.game,
+    required this.victory,
+    super.key,
+  });
+
+  final ct_models.Game game;
+  final ct_models.VictoryState victory;
+
+  @override
+  State<VictoryOverlay> createState() => _VictoryOverlayState();
+}
+
+class _VictoryOverlayState extends State<VictoryOverlay> {
+  bool _dismissed = false;
+
+  @override
+  Widget build(BuildContext context) {
+    if (_dismissed) return const SizedBox.shrink();
+    return Positioned.fill(
+      child: Container(
+        color: Colors.black54,
+        child: Center(
+          child: VictoryPanel(
+            game: widget.game,
+            victory: widget.victory,
+            onViewFinalState: () => setState(() => _dismissed = true),
+          ),
+        ),
+      ),
+    );
+  }
+}
+
+class VictoryPanel extends StatelessWidget {
+  const VictoryPanel({
+    required this.game,
+    required this.victory,
+    this.onViewFinalState,
+    super.key,
+  });
+
+  final ct_models.Game game;
+  final ct_models.VictoryState victory;
+  final VoidCallback? onViewFinalState;
+
+  @override
+  Widget build(BuildContext context) {
+    final winner = game.players.firstWhere(
+      (p) => p.id == victory.winnerPlayerId,
+      orElse: () => game.players.first,
+    );
+    final victoryLabel = switch (victory.type) {
+      ct_models.VictoryType.military => 'Military victory',
+    };
+
+    return Padding(
+      padding: const EdgeInsets.all(24),
+      child: CtPanel(
+        padding: const EdgeInsets.all(24),
+        child: Column(
+          mainAxisSize: MainAxisSize.min,
+          children: [
+            Text(
+              victoryLabel,
+              style: Theme.of(context).textTheme.headlineSmall,
+            ),
+            const SizedBox(height: 12),
+            Text(
+              '${winner.displayName} wins on turn ${victory.turnNumber}.',
+              style: Theme.of(context).textTheme.bodyLarge,
+              textAlign: TextAlign.center,
+            ),
+            const SizedBox(height: 24),
+            Row(
+              mainAxisSize: MainAxisSize.min,
+              children: [
+                CtNinePatchButton(
+                  onPressed: () {
+                    Navigator.of(context).popUntil((route) => route.isFirst);
+                  },
+                  child: const Text('Return to main menu'),
+                ),
+                const SizedBox(width: 12),
+                CtNinePatchButton(
+                  onPressed: () {
+                    onViewFinalState?.call();
+                  },
+                  child: const Text('View final state'),
+                ),
+              ],
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
+
